refactor(mcp): drop the deprecated GitHub Models provider
GitHub Models has been deprecated. CI was still routed to it to save
tokens, and it had already started failing: gpt-4o returned 404 from
models.inference.ai.azure.com during the last run, so every agent mapped
to it (gemini-flash-lite-latest, gemini-pro-latest) silently broke, while
gpt-4o-mini still answered.

Remove the provider and its model mapping, and resolve every model
through Gemini in all environments. getModel() stays as the single
indirection point for model selection, now living next to the Google
provider, so no call site changes.

Note this makes HEY_JARVIS_GOOGLE_API_KEY load-bearing in CI: mcp tests
previously avoided Gemini there and now use it.
