+++ mcp/mastra/utils/providers/ollama-provider.ts
@@ -1,6 +1,6 @@
 import type { LanguageModelV3 } from '@ai-sdk/provider';
 import { createOllama } from 'ai-sdk-ollama';
-import { getModel } from './github-models-provider.js';
+import { getModel } from './google-provider.js';
 
 /**
  * Ollama provider for local LLM inference.
@@ -117,8 +117,7 @@ export const ollamaModel: LanguageModelV3 = ollama(OLLAMA_MODEL);
 
 /**
  * Returns the Ollama model when HEY_JARVIS_OLLAMA_BASE_URL is configured,
- * otherwise falls back to getModel(fallbackGeminiModel). getModel() selects
- * GitHub Models (gpt-4o) in CI or the specified Gemini model in production.
+ * otherwise falls back to the given Gemini model.
  * Use this instead of ollamaModel directly to avoid connection errors when
  * Ollama is not available (e.g. in CI environments).
  * @param fallbackGeminiModel - Gemini model name to use as fallback (default: 'gemini-flash-lite-latest')
